Updated TextFile.txt
This file "TextFile.txt" has been just changed in the "FirstBranch" and not the master branch. Once everyone agress with the changes or updates, we can merge it to the master branch and delete the "FirstBranch". 
First Branch is just a copy of master branch, we make changes to First Branch and if everything works fine, then the changes are merged to master branch and the temporary branch can be deleted later.